requirements.txt
# Core dependencies for resume parsing
instructor>=1.0.0
openai>=1.0.0
pydantic>=2.0.0

# PDF processing libraries
PyMuPDF>=1.23.0

# Standard library enhancements
python-dotenv>=1.0.0

# ChromaDB for vector storage
chromadb>=1.0

# Optional dependencies for enhanced functionality
pandas>=2.0.0
numpy>=1.24.0 